requirements.txt
pandas==2.2.3
python-telegram-bot==13.5
python-dotenv==1.0.1